requirements.txt
# Основная библиотека для Discord API
discord.py>=2.3.0

# Для загрузки переменных окружения из .env файла
python-dotenv>=0.20.0

# Библиотеки для работы с БД и ORM
SQLAlchemy>=2.0.20 # ORM, рекомендуется указывать конкретную минорную версию или использовать ~=
asyncpg>=0.27.0    # Асинхронный драйвер для PostgreSQL
alembic>=1.11.0    # Для миграций базы данных

# psycopg2-binary  # Синхронный драйвер для PostgreSQL, если потребуется для каких-то утилит или старых частей кода.
                   # Для основного асинхронного кода asyncpg является предпочтительным.

# Другие зависимости могут быть добавлены по мере необходимости
# Например, для NLU, работы с API и т.д.
